requirements.txt
bcrypt
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
fastapi-authlib
python-dotenv
jinja2